refactor: move command processing to `cmd` module
Move the command processing code into its own module.

This keeps the `main` module clean of utility code, allowing it to just define what the commands are
and provide a high-level match statement for each of them.

Also remove some text from the documentation of the `perform_update_assessment` function which did
not apply after removing support for different privilege levels.
